Fix problem with unconstrained number of decimal digits for exposure value in info-panel
The photo exposure value will often overrun the focal length value in the info-panel when the number
of decimal digits is more than a few long. By convention, there shouldn't be any decimal digits in the
denominator so this fix rounds the denominator value before display. In addition, this fix increases
the number of decimal digits read from the photo exif metadata to improve the precision of the exposure value
to match what exiftool and other exif viewers display.