Add excludeDir to extension interface
Without this, the mechanism to exclude a directory from indexing is a
bit strange.  I had to use the `scanDirectory.after` hook, then remove
elements from the `output.directories` array.

In order to make the extension hook a bit more developer-friendly,
I also changed the arguments to the `excludeDir` function to be a
dictionary instead of just an array (or tuple?) of strings.